backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Pillow==10.0.0
python-dotenv==1.0.0
werkzeug==2.3.7
